requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
bitsandbytes==0.42.0
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.2.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.20.3
    # via
    #   tokenizers
    #   transformers
hydra==2.5
idna==3.6
    # via requests
jinja2==3.1.3
    # via torch
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   pandas
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.19.3
    # via torch
nvidia-nvjitlink-cu12==12.3.101
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==23.2
    # via
    #   huggingface-hub
    #   transformers
pandas==2.2.1
pillow==10.2.0
pygments==2.17.2
    # via rich
python-dateutil==2.8.2
    # via pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
rich==13.7.0
safetensors==0.4.2
    # via transformers
scipy==1.12.0
    # via bitsandbytes
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tokenizers==0.15.2
    # via transformers
torch==2.2.1
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   transformers
transformers==4.38.1
triton==2.2.0
    # via torch
typing-extensions==4.9.0
    # via
    #   huggingface-hub
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
